requirements.txt
paho-mqtt==2.1.0
google-cloud-pubsub==2.27.2
